move hooks into sub dir; make hook for guessing card height to cleanup code
