+++ frontend/src/utils/hooks/useAutoFocusDialog.tsx
