+++ frontend/src/utils/hooks/useStepObserver.ts
